adapters/credentialmate/__init__.py
"""
CredentialMate Adapter

Connects AI Brain to the CredentialMate repository.

CredentialMate:
- FastAPI backend (Python)
- Next.js frontend (TypeScript)
- PostgreSQL database
- L1 autonomy (stricter, HIPAA compliance)

Location: /Users/tmac/credentialmate
"""

from pathlib import Path
import yaml

from ..base import BaseAdapter, AppContext


class CredentialMateAdapter(BaseAdapter):
    """Adapter for CredentialMate project."""

    def __init__(self):
        config_path = Path(__file__).parent / "config.yaml"
        with open(config_path) as f:
            self.config = yaml.safe_load(f)

    def get_context(self) -> AppContext:
        """Get CredentialMate context for Ralph."""
        return AppContext(
            project_name=self.config["project"]["name"],
            project_path=self.config["project"]["path"],
            language=self.config["project"]["language"],
            lint_command=self.config["commands"]["lint"],
            typecheck_command=self.config["commands"]["typecheck"],
            test_command=self.config["commands"]["test"],
            coverage_command=self.config["commands"]["coverage"],
            source_paths=self.config["paths"]["source"],
            test_paths=self.config["paths"]["tests"],
            coverage_report_path=self.config["paths"]["coverage_report"],
            autonomy_level=self.config["autonomy_level"],
        )

    def run_ralph(self, changed_files: list[str], session_id: str) -> dict:
        """Run Ralph verification for CredentialMate."""
        # TODO: Implement in Phase 0
        # Will invoke ralph.engine.verify() with CredentialMate context
        # Additional HIPAA checks may apply
        raise NotImplementedError("CredentialMate Ralph integration not yet implemented")
